extract Instant Book filter view to separate component
